# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/alerts":{"get":{"tags":["SOC Alerts"],"summary":"Get Alerts","operationId":"get_alerts_alerts_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["SOC Alerts"],"summary":"Create Alert","operationId":"create_alert_alerts_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlertCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/actions/{incident_id}":{"get":{"tags":["SOC Alerts"],"summary":"Get Actions By Incident","operationId":"get_actions_by_incident_actions__incident_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"integer","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users":{"get":{"summary":"Get Users","operationId":"get_users_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Create User","operationId":"create_user_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/incidents":{"get":{"tags":["SOC Alerts"],"summary":"Get Incidents","operationId":"get_incidents_incidents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["SOC Alerts"],"summary":"Get Incidents","operationId":"get_incidents_incidents_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/incidents/{incident_id}":{"get":{"tags":["SOC Alerts"],"summary":"Get Incident","operationId":"get_incident_incidents__incident_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"integer","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["SOC Alerts"],"summary":"Update Incident Status","operationId":"update_incident_status_incidents__incident_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"integer","title":"Incident Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentStatusUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["SOC Alerts"],"summary":"Delete Incident","operationId":"delete_incident_incidents__incident_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"integer","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/password-resets":{"get":{"summary":"Get Password Resets","operationId":"get_password_resets_password_resets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Create Password Reset","operationId":"create_password_reset_password_resets_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/password-resets/verify":{"post":{"summary":"Verify Otp","operationId":"verify_otp_password_resets_verify_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OTPVerify"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/password-resets/{request_id}/approve":{"put":{"summary":"Approve Password Reset","operationId":"approve_password_reset_password_resets__request_id__approve_put","parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"integer","title":"Request Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/it-tickets":{"get":{"tags":["IT Support"],"summary":"Get It Tickets","operationId":"get_it_tickets_it_tickets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["IT Support"],"summary":"Create It Ticket","operationId":"create_it_ticket_it_tickets_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ITTicketCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/it-tickets/{ticket_id}/resolve":{"put":{"tags":["IT Support"],"summary":"Resolve Ticket","operationId":"resolve_ticket_it_tickets__ticket_id__resolve_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"integer","title":"Ticket Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/it-tickets/{ticket_id}/escalate":{"put":{"tags":["IT Support"],"summary":"Escalate Ticket","operationId":"escalate_ticket_it_tickets__ticket_id__escalate_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"integer","title":"Ticket Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/it-tickets/{ticket_id}/feedback":{"put":{"tags":["IT Support"],"summary":"Ticket Feedback","operationId":"ticket_feedback_it_tickets__ticket_id__feedback_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"integer","title":"Ticket Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketFeedback"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/it-tickets/{ticket_id}/close":{"put":{"tags":["IT Support"],"summary":"Close Ticket","operationId":"close_ticket_it_tickets__ticket_id__close_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"integer","title":"Ticket Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ticket-history/{ticket_id}":{"get":{"tags":["IT Support"],"summary":"Get Ticket History","operationId":"get_ticket_history_ticket_history__ticket_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"integer","title":"Ticket Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Me","operationId":"get_me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/assistant/chat":{"post":{"tags":["AI Assistant"],"summary":"Chat","operationId":"chat_assistant_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/training/chat":{"post":{"tags":["Training"],"summary":"Training Chat","operationId":"training_chat_training_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/training/recommendations":{"get":{"tags":["Training"],"summary":"Get Recommendations","operationId":"get_recommendations_training_recommendations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/training-videos":{"get":{"tags":["Training Videos"],"summary":"Get Training Videos","operationId":"get_training_videos_training_videos_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Training Videos"],"summary":"Create Training Video","operationId":"create_training_video_training_videos_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingVideoCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/training-videos/{video_id}":{"put":{"tags":["Training Videos"],"summary":"Update Training Video","operationId":"update_training_video_training_videos__video_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"video_id","in":"path","required":true,"schema":{"type":"integer","title":"Video Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainingVideoUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Training Videos"],"summary":"Delete Training Video","operationId":"delete_training_video_training_videos__video_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"video_id","in":"path","required":true,"schema":{"type":"integer","title":"Video Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/security-updates":{"get":{"tags":["Security Updates"],"summary":"Get Updates","operationId":"get_updates_security_updates_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Security Updates"],"summary":"Create Update","operationId":"create_update_security_updates_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SecurityUpdateCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/security-updates/{update_id}":{"delete":{"tags":["Security Updates"],"summary":"Delete Update","operationId":"delete_update_security_updates__update_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"update_id","in":"path","required":true,"schema":{"type":"integer","title":"Update Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard/summary":{"get":{"tags":["Dashboard"],"summary":"Dashboard Summary","operationId":"dashboard_summary_dashboard_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/dashboard/activity":{"get":{"tags":["Dashboard"],"summary":"Dashboard Activity","operationId":"dashboard_activity_dashboard_activity_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/dashboard/charts":{"get":{"tags":["Dashboard"],"summary":"Dashboard Charts","operationId":"dashboard_charts_dashboard_charts_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AlertCreate":{"properties":{"alert_name":{"type":"string","title":"Alert Name"},"alert_data":{"type":"string","title":"Alert Data"}},"type":"object","required":["alert_name","alert_data"],"title":"AlertCreate"},"ChatRequest":{"properties":{"conversation_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Conversation Id"},"message":{"type":"string","title":"Message"},"confirm_action":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Confirm Action"}},"type":"object","required":["message"],"title":"ChatRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ITTicketCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"}},"type":"object","required":["title","description"],"title":"ITTicketCreate"},"IncidentStatusUpdate":{"properties":{"status":{"type":"string","enum":["Open","In Progress","Resolved","Closed"],"title":"Status"}},"type":"object","required":["status"],"title":"IncidentStatusUpdate"},"OTPVerify":{"properties":{"request_id":{"type":"integer","title":"Request Id"},"otp":{"type":"string","title":"Otp"}},"type":"object","required":["request_id","otp"],"title":"OTPVerify"},"PasswordResetCreate":{"properties":{"employee_id":{"type":"string","title":"Employee Id"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["employee_id","reason"],"title":"PasswordResetCreate"},"SecurityUpdateCreate":{"properties":{"title":{"type":"string","title":"Title"},"message":{"type":"string","title":"Message"},"priority":{"type":"string","title":"Priority"}},"type":"object","required":["title","message","priority"],"title":"SecurityUpdateCreate"},"TicketFeedback":{"properties":{"resolved":{"type":"boolean","title":"Resolved"}},"type":"object","required":["resolved"],"title":"TicketFeedback"},"TrainingChatRequest":{"properties":{"video_title":{"type":"string","title":"Video Title"},"question":{"type":"string","title":"Question"}},"type":"object","required":["video_title","question"],"title":"TrainingChatRequest"},"TrainingVideoCreate":{"properties":{"title":{"type":"string","title":"Title"},"topic":{"type":"string","title":"Topic"},"youtube_url":{"type":"string","title":"Youtube Url"},"description":{"type":"string","title":"Description"}},"type":"object","required":["title","topic","youtube_url","description"],"title":"TrainingVideoCreate"},"TrainingVideoUpdate":{"properties":{"title":{"type":"string","title":"Title"},"topic":{"type":"string","title":"Topic"},"youtube_url":{"type":"string","title":"Youtube Url"},"description":{"type":"string","title":"Description"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["title","topic","youtube_url","description","is_active"],"title":"TrainingVideoUpdate"},"UserCreate":{"properties":{"employee_id":{"type":"string","title":"Employee Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"department":{"type":"string","title":"Department"},"role":{"type":"string","title":"Role"}},"type":"object","required":["employee_id","name","email","password","department","role"],"title":"UserCreate"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}