# backend/schemas/password_reset.py
from pydantic import BaseModel

class PasswordResetCreate(BaseModel):
    employee_id: str
    reason: str

class OTPVerify(BaseModel):
    request_id: int
    otp: str

class PasswordResetStatusUpdate(BaseModel):
    status: str